moved createGdsConfigurationMap to an implict as we did for flattenMap
